draft1/modules/module-2-knowledge-graph/requirements.txt
# KGL Module 2: AI Knowledge Graph Service Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pydantic-settings==2.0.3

# AI & ML Libraries
openai==1.3.7
anthropic==0.3.11
sentence-transformers==2.2.2
transformers==4.35.2

# NLP Processing
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Vector Database
qdrant-client==1.6.4

# Scientific Computing
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# Background Jobs
redis==5.0.1
celery==5.3.4

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
asyncio==3.4.3
python-json-logger==2.0.7